Memory leaks resolved + database defination completed. Error Handling yet to be done

+++ src/headers/generic.h
@@ -20,9 +20,9 @@
 #include <bitset>
 #include <numeric>
 
-typedef int64_t ll;
-typedef std::pair<ll,ll> key_value;
-typedef ll t_key;
-typedef ll t_value;
+typedef unsigned long long ull;
+typedef std::pair<ull, ull> key_value;
+typedef ull t_key;
+typedef ull t_value;
 
 #endif
